# openapi_temp.json
{"openapi":"3.1.0","info":{"title":"Module 5 API","version":"1.0.0"},"paths":{"/api/examples/":{"get":{"tags":["examples"],"summary":"Get Examples","operationId":"get_examples_api_examples__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ExampleResponse"},"type":"array","title":"Response Get Examples Api Examples  Get"}}}}}},"post":{"tags":["examples"],"summary":"Create Example","operationId":"create_example_api_examples__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExampleCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExampleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/examples/{example_id}":{"get":{"tags":["examples"],"summary":"Get Example","operationId":"get_example_api_examples__example_id__get","parameters":[{"name":"example_id","in":"path","required":true,"schema":{"type":"integer","title":"Example Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExampleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["examples"],"summary":"Delete Example","operationId":"delete_example_api_examples__example_id__delete","parameters":[{"name":"example_id","in":"path","required":true,"schema":{"type":"integer","title":"Example Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/register":{"post":{"tags":["auth"],"summary":"Register User","description":"ìë¡ì´ ì¬ì©ìë¥¼ ë±ë¡í©ëë¤.\n\n- ì¬ì©ìëª ì¤ë³µ ê²ì¬\n- ì´ë©ì¼ ì¤ë³µ ê²ì¬\n- ë¹ë°ë²í¸ í´ì±\n- ì¬ì©ì ìì±","operationId":"register_user_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"ì¬ì©ì ë¡ê·¸ì¸ ë° JWT í í° ë°ê¸.\n\n- OAuth2PasswordRequestFormì username íëì ì´ë©ì¼ì ìë ¥ë°ìµëë¤.\n- ì´ë©ì¼ê³¼ ë¹ë°ë²í¸ë¥¼ ê²ì¦í©ëë¤.\n- ì±ê³µ ì JWT ì¡ì¸ì¤ í í°ì ë°íí©ëë¤.\n- ì¤í¨ ì 401 Unauthorized ìë¬ë¥¼ ë°íí©ëë¤.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth"],"summary":"Get Me","description":"íì¬ ë¡ê·¸ì¸í ì¬ì©ìì ì ë³´ë¥¼ ë°íí©ëë¤.\n\nAuthorization í¤ëì Bearer í í°ì´ íìí©ëë¤.\ní í°ì´ ìê±°ë ì í¨íì§ ìì ê²½ì° 401 Unauthorized ìë¬ë¥¼ ë°íí©ëë¤.","operationId":"get_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/widgets":{"get":{"tags":["widgets"],"summary":"Get User Widgets","description":"íì¬ ì¬ì©ìì ëª¨ë  ìì ¯ì ì¡°íí©ëë¤.\n\n- ì¸ì¦ íì\n- user_idë¡ íí°ë§íì¬ ì¬ì©ì ê²©ë¦¬ ë³´ì¥","operationId":"get_user_widgets_api_widgets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WidgetResponse"},"type":"array","title":"Response Get User Widgets Api Widgets Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["widgets"],"summary":"Create Widget","description":"ì ìì ¯ì ìì±í©ëë¤.\n\n- ì¸ì¦ íì\n- user_idë íì¬ ë¡ê·¸ì¸í ì¬ì©ìë¡ ìë ì¤ì \n- configì layoutì JSON ë¬¸ìì´ë¡ ì ì¥","operationId":"create_widget_api_widgets_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WidgetCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WidgetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/widgets/{widget_id}":{"put":{"tags":["widgets"],"summary":"Update Widget","description":"ìì ¯ì ìë°ì´í¸í©ëë¤.\n\n- ì¸ì¦ íì\n- ìì ê¶ ê²ì¦ (ë¤ë¥¸ ì¬ì©ìì ìì ¯ ìì  ë¶ê°)\n- ì ê³µë íëë§ ìë°ì´í¸","operationId":"update_widget_api_widgets__widget_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"widget_id","in":"path","required":true,"schema":{"type":"integer","title":"Widget Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WidgetUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WidgetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["widgets"],"summary":"Delete Widget","description":"ìì ¯ì ì­ì í©ëë¤.\n\n- ì¸ì¦ íì\n- ìì ê¶ ê²ì¦ (ë¤ë¥¸ ì¬ì©ìì ìì ¯ ì­ì  ë¶ê°)","operationId":"delete_widget_api_widgets__widget_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"widget_id","in":"path","required":true,"schema":{"type":"integer","title":"Widget Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/widgets/layout":{"put":{"tags":["widgets"],"summary":"Batch Update Layouts","description":"ì¬ë¬ ìì ¯ì ë ì´ììì ì¼ê´ ìë°ì´í¸í©ëë¤.\n\n- ì¸ì¦ íì\n- ê° ìì ¯ì ìì ê¶ ê²ì¦\n- ì¡´ì¬íì§ ìë ìì ¯ IDë ë¬´ì","operationId":"batch_update_layouts_api_widgets_layout_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LayoutBatchUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WidgetResponse"},"type":"array","title":"Response Batch Update Layouts Api Widgets Layout Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/health":{"get":{"summary":"Health Check","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_auth_login_post"},"ExampleCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name"],"title":"ExampleCreate"},"ExampleResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","name","description","created_at","updated_at"],"title":"ExampleResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LayoutBatchUpdate":{"properties":{"layouts":{"items":{"$ref":"#/components/schemas/WidgetLayoutData"},"type":"array","title":"Layouts","description":"List of widget layouts to update"}},"type":"object","required":["layouts"],"title":"LayoutBatchUpdate","description":"ë ì´ìì ì¼ê´ ìë°ì´í¸ ìì²­"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserCreate":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3,"title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":8,"title":"Password"}},"type":"object","required":["username","email","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"is_active":{"type":"boolean","title":"Is Active"},"is_verified":{"type":"boolean","title":"Is Verified"},"last_login":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Login"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","username","email","is_active","is_verified","last_login","created_at","updated_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WidgetCreate":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name","description":"Widget name"},"type":{"type":"string","title":"Type","description":"Widget type","default":"default"},"config":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Config","description":"Widget configuration"},"layout":{"allOf":[{"$ref":"#/components/schemas/WidgetLayoutData"}],"description":"Widget layout data"}},"type":"object","required":["name","layout"],"title":"WidgetCreate","description":"Widget ìì± ìì²­"},"WidgetLayoutData":{"properties":{"x":{"type":"integer","minimum":0.0,"title":"X","description":"X coordinate (>= 0)"},"y":{"type":"integer","minimum":0.0,"title":"Y","description":"Y coordinate (>= 0)"},"w":{"type":"integer","maximum":6.0,"minimum":1.0,"title":"W","description":"Width (1-6)"},"h":{"type":"integer","minimum":1.0,"title":"H","description":"Height (>= 1)"},"i":{"type":"string","title":"I","description":"Unique widget identifier"}},"type":"object","required":["x","y","w","h","i"],"title":"WidgetLayoutData","description":"React-grid-layout ì¢í ë°ì´í°"},"WidgetResponse":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"name":{"type":"string","title":"Name"},"type":{"type":"string","title":"Type"},"config":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Config"},"layout":{"type":"object","title":"Layout"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","user_id","name","type","config","layout","created_at","updated_at"],"title":"WidgetResponse","description":"Widget ìëµ"},"WidgetUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Name","description":"Widget name"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type","description":"Widget type"},"config":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Config","description":"Widget configuration"},"layout":{"anyOf":[{"$ref":"#/components/schemas/WidgetLayoutData"},{"type":"null"}],"description":"Widget layout data"}},"type":"object","title":"WidgetUpdate","description":"Widget ìë°ì´í¸ ìì²­"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/login"}}}}}}